organize code into sub package rpc/viewblock
