+++ platform/zilliqa/block.go
@@ -3,6 +3,7 @@ package zilliqa
 import (
 	"strconv"
 
+	"github.com/trustwallet/blockatlas/platform/zilliqa/viewblock"
 	"github.com/trustwallet/golibs/types"
 )
 
@@ -21,9 +22,17 @@ func (p *Platform) CurrentBlockNumber() (int64, error) {
 
 func (p *Platform) GetBlockByNumber(num int64) (*types.Block, error) {
 	var normalized []types.Tx
-	txs, err := p.rpcClient.GetTxInBlock(num)
+	var txs []viewblock.Tx
+
+	header, rpcTxs, err := p.rpcClient.GetTxInBlock(num)
 	if err != nil {
 		return nil, err
+	}
+
+	for _, tx := range rpcTxs {
+		if tx := TxFromRpc(tx, header); tx != nil {
+			txs = append(txs, *tx)
+		}
 	}
 
 	for _, srcTx := range txs {
